backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.34
aiosqlite==0.20.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-frontmatter==1.1.0
watchdog==4.0.1
python-multipart==0.0.20
python-dotenv==1.0.1

# ── Testing ──────────────────────────────────────────────────
pytest==8.3.5
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.28.1
